// arxflow.client/src/api/generated/model/createAtivoRequest.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * ArxFlow.Server | v1
 * OpenAPI spec version: 1.0.0
 */

export interface CreateAtivoRequest {
  /** @maxLength 50 */
  codAtivo: string;
  /**
   * @maxLength 50
   * @nullable
   */
  tipoAtivo?: string | null;
  /** @nullable */
  emissorId?: number | null;
  /**
   * @maxLength 50
   * @nullable
   */
  alphaToolsId?: string | null;
  /** @nullable */
  dataVencimento?: string | null;
}